FastAPI/backend-auth/src/.ipynb_checkpoints/requirements-checkpoint.txt
fastapi
uvicorn
google-cloud-firestore
numpy
pandas
pydantic
streamlit